Moved serial node operations out of the serial interface. Finished converting the host object to use the serial object. Finished converting to the new event API, though many event handlers are now broken.
git-svn-id: https://s.snth.net/svn/rlk/thumb@1101 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ src/play.hpp
@@ -28,7 +28,7 @@ namespace mode
         virtual void enter();
         virtual void leave();
 
-        virtual bool timer(double);
+        virtual bool timer(int);
 
         virtual double view(const double *);
         virtual void   draw(const double *);
